added the fish piece. Need to setup the hook on my end.
can compare great states and clear the board. Now I need to make the tool to save the board states to files and read them back

I got the state checking to work! Now we just need to enter in the data and get unlocking to work!

can close and open achievement menu

Achievement system works on a basic level. Now we just need to add in the art assets.